refactor(场景): 按模型与实现分组
