+++ src-tauri/src/scene/model/scene_trait.rs
