+++ src-tauri/src/scene/route_planner.rs
@@ -1,8 +1,10 @@
+//! Graph-only route planning for scene transitions.
+
 use std::collections::{HashMap, VecDeque};
 
 use anyhow::{Result, bail};
 
-use super::{SceneId, 档案库SubSceneId};
+use super::model::{SceneId, 档案库SubSceneId};
 
 #[derive(Debug, PartialEq, Eq)]
 pub(super) struct Route {
