+++ src-tauri/src/scene/scenes/archive.rs
@@ -7,7 +7,9 @@ use std::sync::LazyLock;
 
 use anyhow::Result;
 
-use super::{DEFAULT_THRESHOLD, Scene, SceneAction, SceneId, SceneTransition, 档案库SubSceneId};
+use super::super::{
+    DEFAULT_THRESHOLD, Scene, SceneAction, SceneId, SceneTransition, 档案库SubSceneId,
+};
 use crate::session::Session;
 use crate::utils::region::Region2D;
