+++ src-tauri/src/scene/scenes/mod.rs
@@ -1,0 +1,8 @@
+//! Concrete game scene recognizers and transition definitions.
+
+pub mod archive;
+pub mod overworld;
+pub mod terminal;
+mod unknown;
+
+pub use unknown::Scene未知;
